apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "arewapay-backend"
version = "0.1.0"
description = "ArewaPay API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.14.0",
  "psycopg2-binary>=2.9.10",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt>=4.2.0",
  "slowapi>=0.1.9",
  "httpx>=0.28.0",
  "python-multipart>=0.0.17",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.28.0",
  "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["UP042"]

[tool.ruff.lint.per-file-ignores]
"app/models/**" = ["F821"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
